# audit/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "WB Gamification API",
        "description": "Gamification engine REST API for points, badges, levels, leaderboards, challenges, kudos, and more.",
        "version": "1.6.3",
        "contact": {
            "name": "Wbcom Designs",
            "url": "https://wbcomdesigns.com"
        }
    },
    "servers": [
        {
            "url": "http://wb-gamification.local/wp-json/wb-gamification/v1"
        }
    ],
    "paths": {
...
        "/rules": {
            "get": {
                "summary": "Get rules",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            },
            "post": {
                "summary": "Create rules",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rule_type": {
                                        "type": "string",
                                        "description": ""
                                    },
                                    "target_id": {
                                        "type": "string",
                                        "description": "Badge ID (for badge_condition) or action ID (for points_multiplier)."
                                    },
                                    "rule_config": {
                                        "type": "object",
                                        "description": "JSON condition/multiplier config. Structure depends on rule_type."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "description": ""
                                    }
                                },
                                "required": [
                                    "rule_type",
                                    "rule_config"
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/rules/{id}": {
            "get": {
                "summary": "Get rules",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "post": {
                "summary": "Create rules",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rule_type": {
                                        "type": "string",
                                        "description": ""
                                    },
                                    "target_id": {
                                        "type": "string",
                                        "description": "Badge ID (for badge_condition) or action ID (for points_multiplier)."
                                    },
                                    "rule_config": {
                                        "type": "object",
                                        "description": "JSON condition/multiplier config. Structure depends on rule_type."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "description": ""
                                    }
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "summary": "Update rules",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rule_type": {
                                        "type": "string",
                                        "description": ""
                                    },
                                    "target_id": {
                                        "type": "string",
                                        "description": "Badge ID (for badge_condition) or action ID (for points_multiplier)."
                                    },
                                    "rule_config": {
                                        "type": "object",
                                        "description": "JSON condition/multiplier config. Structure depends on rule_type."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "description": ""
                                    }
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "patch": {
                "summary": "Update rules",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rule_type": {
                                        "type": "string",
                                        "description": ""
                                    },
                                    "target_id": {
                                        "type": "string",
                                        "description": "Badge ID (for badge_condition) or action ID (for points_multiplier)."
                                    },
                                    "rule_config": {
                                        "type": "object",
                                        "description": "JSON condition/multiplier config. Structure depends on rule_type."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "description": ""
                                    }
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "delete": {
                "summary": "Delete rules",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-WP-Nonce",
                "description": "WordPress REST API nonce for cookie-based auth."
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-WB-Gam-Key",
                "description": "API key for cross-site authentication."
            }
        }
    },
    "security": [
        {
            "cookieAuth": []
        },
        {
            "apiKeyAuth": []
        }
    ]
}