# src/API/RulesController.php
<?php
/**
 * REST API: Rules Controller
 *
 * CRUD management for gamification rules stored in wb_gam_rules.
 *
 * GET    /wb-gamification/v1/rules           List rules (filterable by type)
 * POST   /wb-gamification/v1/rules           Create a new rule
 * GET    /wb-gamification/v1/rules/{id}      Get single rule
 * PUT    /wb-gamification/v1/rules/{id}      Update a rule
 * DELETE /wb-gamification/v1/rules/{id}      Delete a rule
 *
 * All endpoints require manage_options.
 *
 * Rule types:
 *   badge_condition  — conditions evaluated by BadgeEngine after point awards
 *     A rule is a GROUP: a match mode over a list of conditions.
 *       { "match": "all", "conditions": [
 *           { "type": "action_count", "action_id": "wp_publish_post", "count": 10 },
 *           { "type": "level_reached", "level_id": 4 }
 *       ] }
 *     The pre-1.6.4 single-condition form is still accepted and is wrapped into a one-condition
 *     group on the way in, so integrators calling it keep working and the database keeps one shape:
 *       { "condition_type": "point_milestone", "points": 500 }
 *
 *   points_multiplier — temporary or permanent per-action point multipliers
 *     Config examples:
 *       { "action_id": "bp_activity_update", "multiplier": 2.0 }
 *       { "multiplier": 1.5, "starts_at": "2026-10-03", "ends_at": "2026-10-04" }
 *     starts_at / ends_at are optional and read in the site time zone; a bare end date covers that
 *     whole day. Outside the window the multiplier does not apply; an unreadable date is a 400.
 *
 * @package WB_Gamification
 * @since   0.1.0
 */

namespace WBGam\API;

use WBGam\Engine\BadgeRule;
use WBGam\Engine\RuleEngine;
use WP_REST_Controller;
use WP_REST_Response;
use WP_REST_Request;
use WP_REST_Server;
use WP_Error;

defined( 'ABSPATH' ) || exit;
// Silencing convention-driven false positives so Plugin Check signal stays clean:
// - PrefixAllGlobals.NonPrefixedHooknameFound — plugin uses `wb_gam_*` as its
// established hook prefix (documented in CLAUDE.md, declared in .phpcs.xml).
// Plugin Check auto-detects `wb_gamification` from the text-domain header
// and doesn't share the .phpcs.xml prefix list; hooks like
// `wb_gam_points_redeemed` are part of the public 1.0 API and can't rename.
// - PrefixAllGlobals.NonPrefixedFunctionFound — same convention. Helper
// functions exported under `wb_gam_*` are documented in `src/Extensions/`.
// - PluginCheck.Security.DirectDB.UnescapedDBParameter +
// WordPress.DB.PreparedSQL.InterpolatedNotPrepared — this file does custom-
// table work. Table names are interpolated from `{$wpdb->prefix}` plus
// literal constants (no user input); user-supplied values pass through
// `$wpdb->prepare()`. MySQL doesn't allow placeholder table names, so the
// interpolation is unavoidable.
// phpcs:disable PluginCheck.Security.DirectDB.UnescapedDBParameter, WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching

/**
 * REST API controller for gamification rules management.
 *
 * Handles CRUD for badge conditions and point multipliers at
 * GET|POST /wb-gamification/v1/rules and GET|PUT|DELETE /wb-gamification/v1/rules/{id}.
 *
 * @package WB_Gamification
 * @since   0.1.0
 */
class RulesController extends WP_REST_Controller {

	/**
	 * REST API namespace.
	 *
	 * @var string
	 */
	protected $namespace = 'wb-gamification/v1';

	/**
	 * REST API route base.
	 *
	 * @var string
	 */
	protected $rest_base = 'rules';

	/**
	 * Allowed rule types.
	 *
	 * @var string[]
	 */
	private const VALID_RULE_TYPES = array( 'badge_condition', 'points_multiplier' );

	/**
	 * Condition types a badge rule may use.
	 *
	 * Used for the error message only -- BadgeRule::sanitize_condition() is what actually decides
	 * what is accepted, and this list exists to tell a caller what they got wrong. Keep them in step.
	 *
	 * @var string[]
	 */
	private const VALID_CONDITION_TYPES = array(
		'point_milestone',
		'points_in_period',
		'action_count',
		'level_reached',
		'badge_earned',
		'streak_days',
		'tenure_days',
	);

	/**
	 * Register REST API routes.
	 *
	 * @return void
	 */
	public function register_routes(): void {
		// Collection.
		register_rest_route(
			$this->namespace,
			'/' . $this->rest_base,
			array(
				array(
					'methods'             => WP_REST_Server::READABLE,
					'callback'            => array( $this, 'get_items' ),
					'permission_callback' => array( $this, 'admin_check' ),
					'args'                => array(
						'rule_type' => array(
							'type' => 'string',
							'enum' => self::VALID_RULE_TYPES,
						),
						'target_id' => array(
							'type'              => 'string',
							'sanitize_callback' => 'sanitize_key',
						),
						'is_active' => array(
							'type' => 'boolean',
						),
					),
				),
				array(
					'methods'             => WP_REST_Server::CREATABLE,
					'callback'            => array( $this, 'create_item' ),
					'permission_callback' => array( $this, 'admin_check' ),
					'args'                => $this->rule_args( required: true ),
				),
				'schema' => array( $this, 'get_item_schema' ),
			)
		);

		// Single item.
		register_rest_route(
			$this->namespace,
			'/' . $this->rest_base . '/(?P<id>[\d]+)',
			array(
				array(
					'methods'             => WP_REST_Server::READABLE,
					'callback'            => array( $this, 'get_item' ),
					'permission_callback' => array( $this, 'admin_check' ),
					'args'                => array(
						'id' => array(
							'type'    => 'integer',
							'minimum' => 1,
						),
					),
				),
				array(
					'methods'             => WP_REST_Server::EDITABLE,
					'callback'            => array( $this, 'update_item' ),
					'permission_callback' => array( $this, 'admin_check' ),
					'args'                => $this->rule_args( required: false ),
				),
				array(
					'methods'             => WP_REST_Server::DELETABLE,
					'callback'            => array( $this, 'delete_item' ),
					'permission_callback' => array( $this, 'admin_check' ),
				),
				'schema' => array( $this, 'get_item_schema' ),
			)
		);
	}

	// ── Callbacks ────────────────────────────────────────────────────────────

	/**
	 * Retrieve a list of gamification rules, optionally filtered by type or target.
	 *
	 * @param WP_REST_Request $request Full details about the request.
	 * @return WP_REST_Response Response containing matching rules.
	 */
	public function get_items( $request ): WP_REST_Response {
		global $wpdb;

		$where  = array( '1=1' );
		$params = array();

		if ( ! empty( $request['rule_type'] ) ) {
			$where[]  = 'rule_type = %s';
			$params[] = $request['rule_type'];
		}

		if ( ! empty( $request['target_id'] ) ) {
			$where[]  = 'target_id = %s';
			$params[] = $request['target_id'];
		}

		if ( null !== $request['is_active'] ) {
			$where[]  = 'is_active = %d';
			$params[] = $request['is_active'] ? 1 : 0;
		}

		$sql = "SELECT id, rule_type, target_id, rule_config, is_active, created_at
		          FROM {$wpdb->prefix}wb_gam_rules
		         WHERE " . implode( ' AND ', $where ) . '
		         ORDER BY rule_type, id ASC';

		// Dynamic WHERE clause; table comes from $wpdb->prefix only; user values go through $wpdb->prepare().
		// phpcs:disable WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.DirectDatabaseQuery.NoCaching
		$rows = $params
			? $wpdb->get_results( $wpdb->prepare( $sql, ...$params ), ARRAY_A )
			: $wpdb->get_results( $sql, ARRAY_A );
		// phpcs:enable WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.DirectDatabaseQuery.NoCaching

		return rest_ensure_response( array_map( array( $this, 'prepare_item' ), $rows ?: array() ) );
	}

	/**
	 * Retrieve a single gamification rule by ID.
	 *
	 * @param WP_REST_Request $request Full details about the request.
	 * @return WP_REST_Response|WP_Error Response on success, WP_Error on failure.
	 */
	public function get_item( $request ): WP_REST_Response|WP_Error {
		$row = $this->fetch_row( (int) $request['id'] );
		return $row
			? rest_ensure_response( $this->prepare_item( $row ) )
			: new WP_Error( 'rest_not_found', __( 'Rule not found.', 'wb-gamification' ), array( 'status' => 404 ) );
	}

	/**
	 * Create a new gamification rule (admin only).
	 *
	 * @param WP_REST_Request $request Full details about the request.
	 * @return WP_REST_Response|WP_Error Response on success, WP_Error on failure.
	 */
	public function create_item( $request ): WP_REST_Response|WP_Error {
		global $wpdb;

		$rule_config = $request['rule_config'];

		// Badge conditions are normalised to the one grouped shape before they are written. What the
		// caller posted is NOT what we store -- storing the raw payload is what would put a second
		// shape back in the database.
		if ( 'badge_condition' === $request['rule_type'] ) {
			$rule_config = $this->normalize_badge_condition_config( $rule_config );
			if ( is_wp_error( $rule_config ) ) {
				return $rule_config;
			}
		}
		$window_error = 'points_multiplier' === $request['rule_type'] ? $this->check_multiplier_window( (array) $rule_config ) : null;
		if ( $window_error ) {
			return $window_error;
		}

		$inserted = $wpdb->insert(
			$wpdb->prefix . 'wb_gam_rules',
			array(
				'rule_type'   => $request['rule_type'],
				'target_id'   => $request['target_id'] ?? null,
				'rule_config' => wp_json_encode( $rule_config ),
				'is_active'   => 1,
				'created_at'  => current_time( 'mysql', true ),
			),
			array( '%s', '%s', '%s', '%d', '%s' )
		);

		if ( ! $inserted ) {
			return new WP_Error( 'rest_insert_failed', __( 'Could not create rule.', 'wb-gamification' ), array( 'status' => 500 ) );
		}

		$this->flush_rules_cache();

		return new WP_REST_Response( $this->prepare_item( $this->fetch_row( $wpdb->insert_id ) ), 201 );
	}

	/**
	 * Update an existing gamification rule (admin only).
	 *
	 * @param WP_REST_Request $request Full details about the request.
	 * @return WP_REST_Response|WP_Error Response on success, WP_Error on failure.
	 */
	public function update_item( $request ): WP_REST_Response|WP_Error {
		global $wpdb;

		$id  = (int) $request['id'];
		$row = $this->fetch_row( $id );
		if ( ! $row ) {
			return new WP_Error( 'rest_not_found', __( 'Rule not found.', 'wb-gamification' ), array( 'status' => 404 ) );
		}

		$data = array();

		if ( isset( $request['rule_type'] ) ) {
			$data['rule_type'] = $request['rule_type'];
		}

		if ( isset( $request['target_id'] ) ) {
			$data['target_id'] = $request['target_id'];
		}

		if ( isset( $request['rule_config'] ) ) {
			$config    = $request['rule_config'];
			$rule_type = $data['rule_type'] ?? $row['rule_type'];
			if ( 'badge_condition' === $rule_type ) {
				$config = $this->normalize_badge_condition_config( $config );
				if ( is_wp_error( $config ) ) {
					return $config;
				}
			}
			$window_error = 'points_multiplier' === $rule_type ? $this->check_multiplier_window( (array) $config ) : null;
			if ( $window_error ) {
				return $window_error;
			}
			$data['rule_config'] = wp_json_encode( $config );
		}

		if ( isset( $request['is_active'] ) ) {
			$data['is_active'] = (int) $request['is_active'];
		}

		if ( empty( $data ) ) {
			return rest_ensure_response( $this->prepare_item( $row ) );
		}

		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.NoCaching -- Write operation; cache flushed via flush_rules_cache().
		$updated = $wpdb->update( $wpdb->prefix . 'wb_gam_rules', $data, array( 'id' => $id ) );
		if ( false === $updated ) {
			return new WP_Error( 'rest_update_failed', __( 'Could not update rule.', 'wb-gamification' ), array( 'status' => 500 ) );
		}
		$this->flush_rules_cache();

		return rest_ensure_response( $this->prepare_item( $this->fetch_row( $id ) ) );
	}

	/**
	 * Delete a gamification rule (admin only).
	 *
	 * @param WP_REST_Request $request Full details about the request.
	 * @return WP_REST_Response|WP_Error Response on success, WP_Error on failure.
	 */
	public function delete_item( $request ): WP_REST_Response|WP_Error {
		global $wpdb;

		$id  = (int) $request['id'];
		$row = $this->fetch_row( $id );
		if ( ! $row ) {
			return new WP_Error( 'rest_not_found', __( 'Rule not found.', 'wb-gamification' ), array( 'status' => 404 ) );
		}

		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.NoCaching -- Write operation; cache flushed via flush_rules_cache().
		$deleted = $wpdb->delete( $wpdb->prefix . 'wb_gam_rules', array( 'id' => $id ), array( '%d' ) );
		if ( false === $deleted ) {
			return new WP_Error( 'rest_delete_failed', __( 'Could not delete rule.', 'wb-gamification' ), array( 'status' => 500 ) );
		}
		$this->flush_rules_cache();

		return new WP_REST_Response(
			array(
				'deleted' => true,
				'id'      => $id,
			),
			200
		);
	}

	/**
	 * Fetch a single rule row from the database by ID.
	 *
	 * @param int $id Rule row ID.
	 * @return array|null Row as associative array, or null if not found.
	 */
	private function fetch_row( int $id ): ?array {
		global $wpdb;
		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.NoCaching -- Admin-only lookup; result used immediately, not suitable for a general cache.
		$row = $wpdb->get_row(
			$wpdb->prepare(
				"SELECT id, rule_type, target_id, rule_config, is_active, created_at
				   FROM {$wpdb->prefix}wb_gam_rules WHERE id = %d",
				$id
			),
			ARRAY_A
		);
		return $row ?: null;
	}

	/**
	 * Shape a raw rule DB row into the REST response format.
	 *
	 * @param array $row Raw row from the rules table.
	 * @return array Formatted rule data for the REST response.
	 */
	private function prepare_item( array $row ): array {
		return array(
			'id'          => (int) $row['id'],
			'rule_type'   => $row['rule_type'],
			'target_id'   => $row['target_id'],
			'rule_config' => json_decode( $row['rule_config'] ?? '{}', true ) ?: array(),
			'is_active'   => (bool) $row['is_active'],
			'created_at'  => $row['created_at'],
		);
	}

	/**
	 * Reject a multiplier campaign window the engine could not read, instead of storing it and
	 * ignoring it. Dates are read in the site's time zone.
	 *
	 * @param array $config The rule_config array from the request.
	 * @return WP_Error|null
	 */
	private function check_multiplier_window( array $config ): ?WP_Error {
		foreach ( array(
			'starts_at' => false,
			'ends_at'   => true,
		) as $key => $is_end ) {
			if ( ! empty( $config[ $key ] ) && null === RuleEngine::window_timestamp( $config[ $key ], $is_end ) ) {
				return new WP_Error(
					'wb_gam_invalid_window',
					/* translators: %s: starts_at or ends_at. */
					sprintf( __( '"%s" must be a date such as 2026-10-04 or 2026-10-04 18:00 (site time).', 'wb-gamification' ), $key ),
					array( 'status' => 400 )
				);
			}
		}
		$start = RuleEngine::window_timestamp( $config['starts_at'] ?? null, false );
		$end   = RuleEngine::window_timestamp( $config['ends_at'] ?? null, true );
		if ( null !== $start && null !== $end && $end <= $start ) {
			return new WP_Error( 'wb_gam_invalid_window', __( '"ends_at" must be after "starts_at".', 'wb-gamification' ), array( 'status' => 400 ) );
		}
		return null;
	}

	/**
	 * Normalise a badge_condition rule_config, or reject it.
	 *
	 * This used to be a validator with its OWN list of three condition types, written before the
	 * badges controller grew a sanitizer that knows all eight. Two validators for one shape is how
	 * the shape drifts: by 1.6.4 this one would reject `level_reached` as invalid and accept only the
	 * pre-1.6.4 `condition_type` vocabulary -- so the documented rules API could no longer create a
	 * rule the engine could actually evaluate.
	 *
	 * There is now one sanitizer, in BadgeRule, and both controllers call it. It still accepts the
	 * `condition_type` form this endpoint documents, and it returns the grouped shape that is the
	 * only thing we write.
	 *
	 * @param array $config The rule_config array from the request.
	 * @return array|WP_Error The grouped config to persist, or WP_Error if nothing valid was posted.
	 */
	private function normalize_badge_condition_config( array $config ): array|WP_Error {
		$group = BadgeRule::from_request( $config );

		if ( null === $group ) {
			return new WP_Error(
				'invalid_condition',
				sprintf(
					/* translators: %s = comma-separated list of valid condition types */
					__( 'A badge_condition rule needs at least one condition. Valid types: %s.', 'wb-gamification' ),
					implode( ', ', self::VALID_CONDITION_TYPES )
				),
				array( 'status' => 400 )
			);
		}

		return $group;
	}

	/**
	 * Bust the rules caches so new/updated rules take effect immediately.
	 */
	private function flush_rules_cache(): void {
		wp_cache_delete( 'wb_gam_badge_rules', 'wb_gamification' );
		wp_cache_delete( 'wb_gam_multiplier_rules', 'wb_gamification' );
	}

	/**
	 * Return REST API argument definitions for rule create/update requests.
	 *
	 * @param bool $required Whether the fields should be required.
	 * @return array Argument definition array.
	 */
	private function rule_args( bool $required = true ): array {
		return array(
			'rule_type'   => array(
				'required'          => $required,
				'type'              => 'string',
				'enum'              => self::VALID_RULE_TYPES,
				'sanitize_callback' => 'sanitize_key',
			),
			'target_id'   => array(
				'type'              => 'string',
				'sanitize_callback' => 'sanitize_key',
				'description'       => 'Badge ID (for badge_condition) or action ID (for points_multiplier).',
			),
			'rule_config' => array(
				'required'    => $required,
				'type'        => 'object',
				'description' => 'JSON condition/multiplier config. Structure depends on rule_type.',
			),
			'is_active'   => array(
				'type'    => 'boolean',
				'default' => true,
			),
		);
	}

	/**
	 * Check if the current user can manage gamification rules.
	 *
	 * Rules govern points-per-action, level thresholds, and challenge
	 * gating — a strictly broader surface than badge CRUD. Tightened
	 * to require its own dedicated cap (wb_gam_manage_rules) so granting
	 * a community manager the badge cap does NOT also expose rules.
	 * Administrators (manage_options) keep working unchanged.
	 *
	 * @return true|WP_Error True if the request has permission, WP_Error otherwise.
	 */
	public function admin_check(): bool|WP_Error {
		if ( \WBGam\Engine\Capabilities::user_can( 'wb_gam_manage_rules' ) ) {
			return true;
		}
		return new WP_Error( 'rest_forbidden', __( 'You do not have permission to manage rules.', 'wb-gamification' ), array( 'status' => 403 ) );
	}

	/**
	 * Retrieve the JSON schema for a rule item.
	 *
	 * @return array JSON schema definition.
	 */
	public function get_item_schema(): array {
		return array(
			'$schema'    => 'http://json-schema.org/draft-04/schema#',
			'title'      => 'wb-gamification-rule',
			'type'       => 'object',
			'properties' => array(
				'id'          => array( 'type' => 'integer' ),
				'rule_type'   => array(
					'type' => 'string',
					'enum' => self::VALID_RULE_TYPES,
				),
				'target_id'   => array( 'type' => array( 'string', 'null' ) ),
				'rule_config' => array( 'type' => 'object' ),
				'is_active'   => array( 'type' => 'boolean' ),
				'created_at'  => array(
					'type'        => 'string',
					'description' => 'UTC, Y-m-d H:i:s.',
				),
			),
		);
	}
}